+++ Frontend/src/app/Config/app.routes.ts
@@ -4,6 +4,11 @@ import { Error404Component } from '../Components/Utils/error404/error404.compone
 import { SignupComponent } from '../Components/Authentication/signup/signup.component';
 import { LoginComponent } from '../Components/Authentication/login/login.component';
 import { DashboardMainComponent } from '../Components/Dashboard/dashboard-main/dashboard-main.component';
+import { FAQComponent } from '../Components/StaticPages/faq/faq.component';
+import { WhatsESNComponent } from '../Components/StaticPages/whats-esn/whats-esn.component';
+import { TheTeamComponent } from '../Components/StaticPages/the-team/the-team.component';
+import { JoinComponent } from '../Components/StaticPages/join/join.component';
+
 
 // Dashboard
 import { MyDashMainComponent } from '../Components/Dashboard/MyDashboard/my-dash-main/my-dash-main.component';
@@ -12,6 +17,7 @@ import { EventsMainComponent } from '../Components/Dashboard/Events/events-main/
 import { AdminMainComponent } from '../Components/Dashboard/Admin/admin-main/admin-main.component';
 import { BureauMainComponent } from '../Components/Dashboard/Bureau/bureau-main/bureau-main.component';
 
+// Guards
 import { AuthGuard } from '../Guards/auth.guard';
 import { AdminGuard } from '../Guards/admin.guard';
 import { BureauGuard } from '../Guards/bureau.guard';
@@ -34,6 +40,11 @@ export const routes: Routes = [
             {path: 'bureau', component: BureauMainComponent, canActivate:[BureauGuard]}
         ]
     },
+    
+    {path:'faq', component:FAQComponent},
+    {path:'esn', component:WhatsESNComponent},
+    {path:'JoinUs', component:JoinComponent},
+    {path:'TheTeam', component:TheTeamComponent},